Initial import of the III-IV code
III-IV is a rudimentary and very opinionated framework to write Rust
web services with.  This first commit is based on almost-verbatim code
I had previously written for a minimal web analytics service.  As is,
this code has been useful for the implementation of another service,
so it makes sense for it to stand on its own.

The name III-IV refers to the number of layers that services using
this framework need to implement.  The 3 is about the rest, driver,
and db layers, and the 4 is about the cross-layer model module.

The main differences in this version compared to the original code I
factored out are that this is split in multiple crates, and there is
a full end-to-end example to demonstrate how this works.  Many of the
interfaces are still pretty ad-hoc right now though, as they come from
just one service.