Add basic authentication features
This extracts code from the EndBASIC Service used to implement
authentication that can stand on its own because it is not tied to any
specific service.

It would be nice to make the functionality here play better with axum,
leveraging extractors, but this will come at a later stage.  For now,
I'm keeping most things as they were.