Move get_required_var to core
This adds a new "env" module in the core crate to capture env-related
functionality, which is necessary to configure various components of
the framework.  Right now this is only get_required_var, but more will
come.

While doing this, parameterize get_required_var so that it can perform
type conversions and thus encapsulate validation in a single place.