Add a basic authn signup/login flow
This is an almost direct copy/paste of the code that exists in the
EndBASIC Service but with all service-specific details removed.

A lot of this needs to be rewritten for cleanliness and to support
proper integration/customization by downstream services.  I've added
a bunch of TODOs for obvious things that need to be looked at, but
they are not exhaustive.