Extend OneShotBuilder to deal with forms
This adds a new `with_query` method that extends the URI in a
OneShotBuilder with a query, a `send_form` method that sends
the content of a form, and a `test_payload_must_be_form` macro
to verify the behavior of a POST handler that expects a form.

While here, also add a `with_fragment`.