Add setup_attach to sqlite and postgres
This extends the sqlite and postgres modules with a new setup_attach
function that can be used to attach a transaction type to an existing
database instance backed by a different transaction type.

This allows implementing various services within the same process, all
sharing the same connection to the database pool.