Include problem message in authorization errors
Fix the formatted content of the RestError::Authorization error to
include the actual problem that was found instead of incorrectly
showing the expected scheme.

This problem existed due to the use of unnamed enum fields because
the string formatter for the error picked the wrong positional field.
So also fix this by using named arguments.