Remove the SmtpMessenger type
This type was first added so that a service could provide a trait with
high-level primitives to send email messages declaratively.  The idea
was that services could then supply a mock messenger that captured the
messages based on their "type".

This worked mostly OK but was overly complex.  In particular, this became
problematic when trying to generalize authentication functionality as
such code needs to send an email but the services cannot implement a
trait for a type when they don't own either of them.

It's easier if we remove the whole idea of the "messenger" and instead
make the services deal with the underlying "mailer" directly.  And to
support this change, this commit provides new test utilities to extract
messages captured from the mock recorder mailer.