Add the EmailTemplate type
This new type can be used to generate email messages based on a
template that contains placeholder strings.  It is more ergonomic
than format!() calls and also allows parameterizing email messages,
which is now necessary to add a generic authentication service.

To support this type, this change also adds a new "template" module
to the core crate which provides a trivial templating engine.  It
will be useful to parameterize HTML responses as well.