Fold the sqlite and postgres crates into core
Separating the sqlite and postgres features into separate crates makes
it difficult to expose a unified interface from the core library.
Furthermore, other crates in the tree, such as the queue crate and the
upcoming changes to the authn crate, end up having to expose features
of their own to select between the two databases... so keeping their
implementations in separate crates wasn't as nice as it sounded.

As a result, merge the two crates into the core and expose them via
features.  This makes the core crate even more complex feature-wise
but I think it'll simplify things overall.