Redo the database layer from the ground up
This gets rid of the whole "typed transactions" idea along with static
dispatch and replaces the architecture with a simpler Db type that can
be wrapped in a Box<dyn Db> and passed around.

Users of this type can obtain an Executor on the Db or on a transaction,
and the Executor provides typed access to the inner PostgreSQL or SQLite
sqlx executors that the services can use to implement their queries.

Other than this being much simpler to understand, this also offers
runtime advantages as can be seen in the queue code: callers can now
choose whether they need to use transactions or not whereas the previous
design forced all callers to use transactions even for standalone
queries.