Pass the DB executor to queue client ops
Instead of holding a reference to the database within the queue client,
pass references to an executor to each operation.

This is to allow the caller more flexibility in deciding whether
enqueuing a task should be part of a transaction or not.  Additionally,
this should also fix some issues that arise when writing tests against
SQLite that have to deal with the database for a service's operations
and also enqueue tasks.