Detect pool timeout errors in postgres tests
For a long time, I've occasionally observed "pool timed out" errors
while tests against PostgreSQL in GitHub.  I had thought this was
due to bugs in the connection retry logic... but as it turns out,
this seems to be just because I forgot treating one specific error
code as retryable.  Fix that now.