Make Db::ex() return a pool connection
Instead of returning a copy of the pool whenever Db::ex() is invoked,
establish a connection and return it.

This is conceptually the right thing to do, as we do not want clients
using more than one connection without noticing, and allow for better
control of retries if the PostgreSQL database has reached its max
number of connections.