Make queue wait operations receive a database
The wait operations perform a loop over multiple queries on the
database so they cannot hold onto an open executor.  Otherwise,
they keep a connection or transaction open, preventing the caller
from issuing other operations.  This is mostly a problem for
PostgreSQL tests where we limit them to just one connection.

Fix this by receiving the database itself and opening new connections
on each iteration of the loop.