Invalidate cached sessions on logout
This is not strictly necessary and it won't do the right thing when
running in Azure Functions (because we don't know how many frontends
the runtime will spawn and we cannot reliably do cache invalidation
across them all), but it's still the right thing to do.  After all,
consumers of this code could run in a single server, and in that
case we would want this to happen.