Add a geolocator based by the IP-API free tier
Because of the limitations of the free tier, this new geolocator is
backed by another geolocator that can provide answers when the free
tier quota is exhausted.

To make this possible, this change adds a new SettableClock fake
clock for testing purposes, which is much nicer to use than the older
MonotonicClock.

Additionally, this change also adds a RequestCounter to keep track of
how many requests have been sent.  I'm keeping this in the geo crate
for now, but this type of request counter should likely be placed in
the core crate; will do so when the need arises.