Rename the lettre crate to smtp
I'm not sure why I thought that naming the crate after its dependency
was a good idea in the first place, given that the use of lettre is
an implementation detail.

The abstractions in this create already use the Smtp prefix instead
of Lettre, so renaming the crate seems fitting.