Return retriable tasks from get_result{,s_since}
Modify the queue's get_result() and get_results_since() primitives to
return tasks that failed on an earlier run and are scheduled to rerun
later.

This keeps all callers the same by making the client's wait() and
wait_all() methods ignore retriable tasks until they are fully done.
However, this adds a new wait_once() to wait for a task until it asks
to retry or finishes, whichever happens first.