Workaround Azure Maps 401 responses
Sometimes, Azure geolocation tests in pull requests fail with a 401,
saying that the Maps key is wrong.  However, this only happens to a
test within the same PR run and only randomly, so the configuration
is correct.

Try to workaround this with retries, which are also a good idea to
have because we are making the tests depend on an external service.